+++ CHANGELOG.md
@@ -1,5 +1,26 @@
 # Release Notes
 
+## [Unreleased](https://github.com/laravel/socialite/compare/v4.1.3...4.0)
+
+
+## [v4.1.3 (2019-04-02)](https://github.com/laravel/socialite/compare/v4.1.2...v4.1.3)
+
+### Fixed
+- Fix bug with no LinkedIn email addresses ([#355](https://github.com/laravel/socialite/pull/355))
+
+
+## [v4.1.2 (2019-03-15)](https://github.com/laravel/socialite/compare/v4.1.1...v4.1.2)
+
+### Fixed
+- Use proper key name for original avatar in Google Provider ([5ec0024](https://github.com/laravel/socialite/commit/5ec0024284d15df527376ced59b9e7b393f6f88b))
+
+
+## [v4.1.1 (2019-03-12)](https://github.com/laravel/socialite/compare/v4.1.0...v4.1.1)
+
+### Fixed
+- Update Google API urls ([#346](https://github.com/laravel/socialite/pull/346))
+
+
 ## [v4.1.0 (2019-02-14)](https://github.com/laravel/socialite/compare/v4.0.3...v4.1.0)
 
 ### Added
@@ -8,20 +29,24 @@
 ### Fixed
 - Handle Google responses excluding 'name' ([#340](https://github.com/laravel/socialite/pull/340))
 
+
 ## [v4.0.3 (2019-01-21)](https://github.com/laravel/socialite/compare/v4.0.2...v4.0.3)
 
 ### Fixed
 - Use proper localized name ([#329](https://github.com/laravel/socialite/pull/329))
+
 
 ## [v4.0.2 (2019-01-21)](https://github.com/laravel/socialite/compare/v4.0.1...v4.0.2)
 
 ### Fixed
 - Fix a bug with no LinkedIn image ([81adfcc](https://github.com/laravel/socialite/commit/81adfcc4f7df3a470cdab8a500db77c0de5d01a3))
 
+
 ## [v4.0.1 (2018-12-20)](https://github.com/laravel/socialite/compare/v4.0.0...v4.0.1)
 
 ### Fixed
 - Fixed broken `GoogleProvider` ([#316](https://github.com/laravel/socialite/pull/316))
+
 
 ## [v4.0.0 (2018-12-18)](https://github.com/laravel/socialite/compare/v3.2.0...v4.0.0)